+++ examples/stochastic_processes.rs
@@ -37,16 +37,16 @@ fn main() {
     let fbm_out = fbm.euler_maruyama(0.0, 0.0, 0.5, 1000, 1, false);
 
     // Plot the paths.
-    plot_vector(abm_out.paths[0].clone(), "./images/arithmetic_brownian_motion.png").unwrap();
-    plot_vector(bdt_out.paths[0].clone(), "./images/black_derman_toy.png").unwrap();
-    plot_vector(bm_out.paths[0].clone(),  "./images/brownian_motion.png").unwrap();
-    plot_vector(cir_out.paths[0].clone(), "./images/cox_ingersoll_ross.png").unwrap();
-    plot_vector(ev_out.paths[0].clone(),  "./images/extended_vasicek.png").unwrap();
-    plot_vector(gbm_out.paths[0].clone(), "./images/geometric_brownian_motion.png").unwrap();
-    plot_vector(hl_out.paths[0].clone(),  "./images/ho_lee.png").unwrap();
-    plot_vector(hw_out.paths[0].clone(),  "./images/hull_white.png").unwrap();
-    plot_vector(ou_out.paths[0].clone(),  "./images/ornstein_uhlenbeck.png").unwrap();
-    plot_vector(fbm_out.paths[0].clone(), "./images/fractional_brownian_motion.png").unwrap();
+    plot_vector!(abm_out.paths[0].clone(), "./images/arithmetic_brownian_motion.png");
+    plot_vector!(bdt_out.paths[0].clone(), "./images/black_derman_toy.png");
+    plot_vector!(bm_out.paths[0].clone(),  "./images/brownian_motion.png");
+    plot_vector!(cir_out.paths[0].clone(), "./images/cox_ingersoll_ross.png");
+    plot_vector!(ev_out.paths[0].clone(),  "./images/extended_vasicek.png");
+    plot_vector!(gbm_out.paths[0].clone(), "./images/geometric_brownian_motion.png");
+    plot_vector!(hl_out.paths[0].clone(),  "./images/ho_lee.png");
+    plot_vector!(hw_out.paths[0].clone(),  "./images/hull_white.png");
+    plot_vector!(ou_out.paths[0].clone(),  "./images/ornstein_uhlenbeck.png");
+    plot_vector!(fbm_out.paths[0].clone(), "./images/fractional_brownian_motion.png");
 }
 
 fn theta_t(_t: f64) -> f64 {